+++ IProblems.cs
@@ -5,7 +5,7 @@
         public void Excute(IOperations operations);
         public void Printing(int i);
         public string GetProblems(int i);
-        public int GetSolutions(int i);
+        public long GetSolutions(int i);
         public int GetProblemsNumber();
 
     }
